+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ NVIDIA Model Optimizer Changelog (Linux)
 **New Features**
 
 - Add support for Transformer Engine quantization for Megatron Core models.
+- Add support for dynamically linked TensorRT plugins in the ONNX quantization workflow.
 
 0.40 (2025-12-12)
 ^^^^^^^^^^^^^^^^^